Refactor error handling in product and review stores: replace logical OR with nullish coalescing for improved error message clarity; update ProductSize component to use nullish coalescing for outOfStock checks; create interface file for Nav components to enhance code organization.
