+++ components/organisms/ProductReview/useFetchReviewsStore.ts
@@ -82,7 +82,7 @@ const useFetchReviewsStore = create<FetchReviewsStore>((set, get) => ({
     } catch (error) {
       if (axios.isAxiosError(error)) {
         set({
-          error: new Error(error.response?.data.message || error.message),
+          error: new Error(error.response?.data.message ?? error.message),
         });
       } else {
         set({
